requirements.txt
click
PyYAML==3.11
paramiko==2.4.2